some styling work
